requirements.txt
# 크롤링
playwright==1.40.0
playwright-stealth==1.0.6

# 비동기 처리
asyncio==3.4.3

# Supabase
supabase==2.3.0

# OpenAI
openai==1.6.1

# 환경변수
python-dotenv==1.0.0

# 로깅
colorlog==6.8.0

# 유틸리티
python-dateutil==2.8.2
pytz==2023.3

# 테스트
pytest==7.4.4
pytest-asyncio==0.23.2

# 개발 도구
black==23.12.1
flake8==7.0.0
isort==5.13.2

# 웹 프레임워크
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 템플릿 엔진
jinja2==3.1.3

# 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 데이터베이스
asyncpg==0.29.0
sqlalchemy==2.0.25
pymysql==1.1.0

# 유효성 검사
pydantic==2.5.3
email-validator==2.1.0

# HTTP 클라이언트
httpx==0.26.0

# 기타
python-json-logger==2.0.7

# 암호화
cryptography==41.0.7
